+++ client/src/index.css
@@ -106,7 +106,7 @@
   --clr-accent-valid: #36ca3b;
   --clr-accent-valid-100: #318b3b;
 
-  --clr-accent-invalid: #ee4f4f;
+  --clr-accent-invalid: #bd0000;
   --clr-accent-invalid-100: #b81c1c;
 
   /* Font Size */
