+++ server/src/controllers/users.controllers.js
@@ -1,6 +1,53 @@
 const { UsersService } = require("../services/users.service");
 
 class UsersController {
+
+  static async registerUser(req, res) {
+    try {
+      const {
+        username,
+        firstname,
+        lastname,
+        email,
+        phone_number,
+        age,
+        address,
+        password,
+      } = req.body;
+      const user = await UsersService.registerUser({
+        username,
+        firstname,
+        lastname,
+        email,
+        phone_number,
+        age,
+        address,
+        password,
+      });
+      res.json(user);
+    } catch (e) {
+      res.status(400).json({ message: "Error creating user" });
+    }
+  }
+
+  static async loginUser(req, res) {
+    try {
+      const {
+        username,
+        email,
+        password,
+      } = req.body;
+      const user = await UsersService.loginUser({
+        username,
+        email,
+        password,
+      });
+      res.json(user);
+    } catch (e) {
+      res.status(400).json({ message: "Error creating user" });
+    }
+  }
+
   static async createUser(req, res) {
     try {
       const {
