--- server/src/routes/registrationRoutes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const { RegistrationController } = require("../controllers/registration.controllers");
-const router = express.Router();
-
-router.post("/", RegistrationController.registerUser);
-
-module.exports = router;
