+++ server/src/routes/userRoutes.js
@@ -8,4 +8,7 @@ router.get("/:id", UsersController.getUserById);
 router.put("/:id", UsersController.updateUserById);
 router.delete("/:id", UsersController.deleteUserById);
 
+router.post("/register", UsersController.registerUser);
+router.post("/login", UsersController.loginUser);
+
 module.exports = router;
